requirements.lock.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
akracer==0.0.14
akshare==1.18.11
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
asyncer==0.0.12
attrs==25.4.0
audioop-lts==0.2.2
backoff==2.2.1
backtrader==1.9.78.123
beautifulsoup4==4.14.3
bidict==0.23.1
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
chainlit==2.9.5
charset-normalizer==3.4.4
chevron==0.14.0
chromadb==0.3.23
click==8.3.1
clickhouse-connect==0.10.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cuid==0.4
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.4.3
eodhd==1.0.32
et_xmlfile==2.0.0
fastapi==0.128.0
feedparser==6.0.12
filelock==3.20.3
filetype==1.2.0
finnhub-python==2.4.26
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.1.0
google-auth==2.47.0
google-genai==1.57.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
hnswlib==0.8.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jsonpatch==1.33
jsonpath==0.82.2
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain-anthropic==1.3.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-experimental==0.4.1
langchain-google-genai==4.2.0
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.2
Lazify==0.4.0
literalai==0.1.201
lxml==6.0.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mcp==1.25.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.15.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-agno==0.50.1
opentelemetry-instrumentation-alephalpha==0.50.1
opentelemetry-instrumentation-anthropic==0.50.1
opentelemetry-instrumentation-bedrock==0.50.1
opentelemetry-instrumentation-chromadb==0.50.1
opentelemetry-instrumentation-cohere==0.50.1
opentelemetry-instrumentation-crewai==0.50.1
opentelemetry-instrumentation-google-generativeai==0.50.1
opentelemetry-instrumentation-groq==0.50.1
opentelemetry-instrumentation-haystack==0.50.1
opentelemetry-instrumentation-lancedb==0.50.1
opentelemetry-instrumentation-langchain==0.50.1
opentelemetry-instrumentation-llamaindex==0.50.1
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-marqo==0.50.1
opentelemetry-instrumentation-mcp==0.50.1
opentelemetry-instrumentation-milvus==0.50.1
opentelemetry-instrumentation-mistralai==0.50.1
opentelemetry-instrumentation-ollama==0.50.1
opentelemetry-instrumentation-openai==0.50.1
opentelemetry-instrumentation-openai-agents==0.50.1
opentelemetry-instrumentation-pinecone==0.50.1
opentelemetry-instrumentation-qdrant==0.50.1
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-replicate==0.50.1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sagemaker==0.50.1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-together==0.50.1
opentelemetry-instrumentation-transformers==0.50.1
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-vertexai==0.50.1
opentelemetry-instrumentation-watsonx==0.50.1
opentelemetry-instrumentation-weaviate==0.50.1
opentelemetry-instrumentation-writer==0.50.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
parsel==1.10.0
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
posthog==7.5.1
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
py-mini-racer==0.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.21
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
questionary==2.1.1
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
sse-starlette==3.1.2
starlette==0.50.0
stockstats==0.6.5
sympy==1.14.0
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.9.1
tqdm==4.67.1
traceloop-sdk==0.50.1
transformers==4.57.5
triton==3.5.1
tushare==1.4.24
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
update-checker==0.18.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
uvloop==0.22.1
w3lib==2.3.1
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.3.2
xlrd==2.0.2
xxhash==3.6.0
yarl==1.22.0
yfinance==1.0
zipp==3.23.0
zstandard==0.25.0
